backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
python-dotenv==1.0.0
django-cors-headers==4.1.0
PyJWT==2.7.0
requests==2.31.0
djangorestframework-simplejwt==5.3.0

# Optional dependencies for MySQL
# mysqlclient==2.1.1  # Uncomment when using MySQL 